runner/pyproject.toml
[project]
name = "runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tim Garvin", email = "tcgarvin@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "pydantic>=2.0.0",
    "structlog>=25.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
]
